multi-routes-ors/server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
geodesic